rename package and localize variable
