+++ app/src/main/java/ca/ualberta/CMPUT3012019T02/alexandria/fragment/exchange/AcceptedFragment.java
@@ -1,4 +1,4 @@
-package ca.ualberta.CMPUT3012019T02.alexandria.fragment.Exchange;
+package ca.ualberta.CMPUT3012019T02.alexandria.fragment.exchange;
 
 import android.os.Bundle;
 import android.support.annotation.NonNull;
@@ -11,14 +11,14 @@ import android.view.ViewGroup;
 import ca.ualberta.CMPUT3012019T02.alexandria.R;
 
 /**
- * Fragment for filtering book list that has the status of Requested
+ * Fragment for filtering book list that has the status of Accepted
  */
 
-public class RequestedFragment extends Fragment {
+public class AcceptedFragment extends Fragment {
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater,
                              @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_exchange_requested,null);
+        return inflater.inflate(R.layout.fragment_exchange_accepted,null);
     }
 }
